ai-model-foundation/requirements.txt
# AI Model Foundation
pyyaml>=6.0
pandas>=2.0
scikit-learn>=1.3
joblib>=1.3
# optional: mlflow for registry backend
